backend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
openai==1.40.0
google-api-python-client==2.138.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0.post1
textract==1.6.5
ultralytics==8.2.65